# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: DevLink KE API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      parameters:
        - name: skill
          in: query
          schema:
            type: string
        - name: level
          in: query
          schema:
            type: string
            enum: [beginner, intermediate, pro]
        - name: location
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /users/top-builders:
    get:
      operationId: getTopBuilders
      tags: [users]
      summary: Top active builders on the platform
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Top builders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        skills:
          type: array
          items:
            $ref: "#/components/schemas/UserSkill"
        level:
          type: string
          enum: [beginner, intermediate, pro]
        githubUrl:
          type: ["string", "null"]
        twitterUrl:
          type: ["string", "null"]
        lookingFor:
          type: ["string", "null"]
        roomsJoined:
          type: integer
        postsCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, username, displayName, skills, level, roomsJoined, postsCount, createdAt]
    UserSkill:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
          enum: [backend, frontend, mobile, ai, networking, design, devops, other]
        proficiency:
          type: string
          enum: [beginner, intermediate, pro]
      required: [name, category, proficiency]
    UserInput:
      type: object
      properties:
        username:
          type: string
        displayName:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
        location:
          type: string
        skills:
          type: array
          items:
            $ref: "#/components/schemas/UserSkill"
        level:
          type: string
          enum: [beginner, intermediate, pro]
        githubUrl:
          type: string
        twitterUrl:
          type: string
        lookingFor:
          type: string
      required: [username, displayName, skills, level]
    UserUpdate:
      type: object
      properties:
        displayName:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
        location:
          type: string
        skills:
          type: array
          items:
            $ref: "#/components/schemas/UserSkill"
        level:
          type: string
          enum: [beginner, intermediate, pro]
        githubUrl:
          type: string
        twitterUrl:
          type: string
        lookingFor:
          type: string
...