# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DevLink KE API
 * OpenAPI spec version: 0.1.0
 */
import type { UserSkill } from './userSkill';
import type { UserUpdateLevel } from './userUpdateLevel';

export interface UserUpdate {
  displayName?: string;
  bio?: string;
  avatarUrl?: string;
  location?: string;
  skills?: UserSkill[];
  level?: UserUpdateLevel;
  githubUrl?: string;
  twitterUrl?: string;
  lookingFor?: string;
}